Use a git submodule to include A-Painter.
The server uses AFRAME.registerSystem() to extend the A-Painter code
that it serves up so that it shares headset, controller and stroke
data.  If the A-Painter or upstream A-Frame code changes, it isn't
always compatible with what the server does.  Using a submodule
simplifies deployment for users by using a version of the A-Painter
code known to be compatible with the server.

Updated the instructions in README.md to match the changes, provide
port/firewall tips and fix some typos.